make DEFTEST not accept any number of arguments
When a DEFTEST function is called, it records the argument list in its
TRIAL for TRY::@RERUN, or at least its best reconstruction of it.
Previously, the code that handled this added a &REST argument to the
lambda list of the DEFTEST, which defeated argument checking, and it
wasn't able to reconstruct the real arglist either.

The vast majority of DEFTESTs have no arguments, so the arglist is
trivially (). However, correct and fast arglist reconstruction in the
general case is non-trivial. Thanks to Gleefre on Libera #commonlisp
who came up with a nice solution.

The lambda list may still be  changed (by adding &REST before &KEY, or
adding SUPPLIEDP to &OPTIONAL arguments), but it's compatible with
the original.